--- Assets/InteractTest.cs
@@ -1,21 +1,0 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public class InteractTest : Interactable {
-
-	// Use this for initialization
-	void Start () {
-		
-	}
-	
-	// Update is called once per frame
-	void Update () {
-		
-	}
-
-    public override void Interact()
-    {
-        Debug.Log("Wollah!");
-    }
-}
