--- Assets/InteractTest.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: 5fb71ab2e894ebc4f8b8b6ab363124f4
-timeCreated: 1512734466
-licenseType: Free
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
