--- Assets/Interactable.cs
@@ -1,7 +1,0 @@
-﻿using UnityEngine;
-using System.Collections;
-
-public abstract class Interactable : MonoBehaviour
-{
-    public abstract void Interact();
-}
