+++ Assets/Scripts.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: d6e2b48c94f903d438da720a60453a81
+folderAsset: yes
+timeCreated: 1512737125
+licenseType: Free
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
