+++ Assets/Scripts/Blocks/Interactable.cs
@@ -1,0 +1,7 @@
+﻿using UnityEngine;
+using System.Collections;
+
+public abstract class Interactable : MonoBehaviour
+{
+    public abstract void Interact();
+}
