+++ Assets/Scripts/Blocks/Interactable.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 2d08f49eb8a925043bd02c80d7013e8a
+timeCreated: 1512734459
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
