+++ Assets/Scripts/Blocks/MinusBlock.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 72619a354eb4aa943a33b7a02f907493
+timeCreated: 1512737239
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
