--- Assets/blobPlacer.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: 4387bf540cdae1e468a2bc576646f32b
-timeCreated: 1512654424
-licenseType: Free
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
