+++ Assets/blockMat.mat.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: abba88489a4b9c6428a5053c53afa67c
+timeCreated: 1512738415
+licenseType: Free
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
